Upgrade lorem-ipsum package; create initial DataList component and wire into existing component demos; create a common lipsum generator function.
